feat: Write createNavBar function & function call
The createNavBar function loops through all sections in index.html
and creates a list item & anchor item for each element. The text of the
anchor item is set to equal the value in the data-nav attribute. The
href attribute is set to equal the id of the corresponding section.

A call to createNavBar dynamically builds the Nav Bar in Javascript